更新leetcode 441-443, 445, 447-457, 459

+++ cpp/leetcode/447. number-of-boomerangs.cpp
@@ -1,9 +1,9 @@
 #include "public.h"
 
-//824ms, 52.35%
-//半暴力
-//定中间的点, 然后求其他点和它距离的平方, 统计频率
-//用大数据long long, 防止溢出
+//568ms, 90.39%
+//暴力O(n^2)
+//定中间的点, 然后用unordered_map记录其他点和它距离的平方的次数
+//注: 提交结果200ms左右似乎有vector记录次数的方法, 时间效率更高, 但代码复杂度很大
 
 class Solution {
 public:
@@ -12,18 +12,14 @@ public:
 
 		for (auto& middle : points)
 		{
-			unordered_map<long long, int> distance_squares;
-
+			unordered_map<int, int> distance_squares;
 			for (auto& side : points)
 			{
-				distance_squares[((long long)middle[0] - side[0])*((long long)middle[0] - side[0]) + ((long long)middle[1] - side[1])*((long long)middle[1] - side[1])]++;
+				distance_squares[(middle[0] - side[0])*(middle[0] - side[0]) + (middle[1] - side[1])*(middle[1] - side[1])]++;
 			}
-
 			for (auto& distance_square : distance_squares)
 				res += (distance_square.second - 1)*distance_square.second;
 		}
-
 		return res;
 	}
 };
-
